use NonEmpty for flowInline
