+++ daemon/src/GHCSpecter/Render/Components/Util.hs
@@ -10,6 +10,7 @@ module GHCSpecter.Render.Components.Util (
 
 import Data.List qualified as L
 import Data.List.NonEmpty (NonEmpty)
+import Data.List.NonEmpty qualified as NE
 import GHCSpecter.Graphics.DSL (
   DrawText (..),
   Polyline (..),
@@ -18,6 +19,7 @@ import GHCSpecter.Graphics.DSL (
   Shape (..),
   ViewPort (..),
   viewPortHeight,
+  viewPortSum,
   viewPortWidth,
  )
 
@@ -45,15 +47,6 @@ moveBoundingBoxBy :: (Double, Double) -> ViewPort -> ViewPort
 moveBoundingBoxBy (dx, dy) (ViewPort (vx0, vy0) (vx1, vy1)) =
   ViewPort (vx0 + dx, vy0 + dy) (vx1 + dx, vy1 + dy)
 
-viewPortSum :: Maybe ViewPort -> ViewPort -> ViewPort
-viewPortSum Nothing vp = vp
-viewPortSum (Just (ViewPort (x0, y0) (x1, y1))) (ViewPort (x0', y0') (x1', y1')) =
-  let x0'' = min x0 x0'
-      y0'' = min y0 y0'
-      x1'' = max x1 x1'
-      y1'' = max y1 y1'
-   in ViewPort (x0'', y0'') (x1'', y1'')
-
 getLeastUpperBoundingBox :: NonEmpty (Primitive e) -> ViewPort
 getLeastUpperBoundingBox itms =
   let vps = fmap primBoundingBox itms
@@ -64,28 +57,34 @@ getLeastUpperBoundingBox itms =
 toSizedLine :: NonEmpty (Primitive a) -> (ViewPort, NonEmpty (Primitive a))
 toSizedLine xs = (getLeastUpperBoundingBox xs, xs)
 
+-- for now
+viewPortSum' = viewPortSum . Just
+
 -- | place grouped items horizontally
 flowInline ::
   -- | initial x offset
   Double ->
   -- | rendered items grouped by each x
-  [NonEmpty (Primitive e)] ->
+  NonEmpty (NonEmpty (Primitive e)) ->
   -- | (bounding box of the collection, placed items)
-  (Maybe ViewPort, [NonEmpty (Primitive e)])
-flowInline offset0 itms0 = (mvp1, itms1)
+  (ViewPort, NonEmpty (NonEmpty (Primitive e)))
+flowInline offset0 itmss0 = (vp1, itmss1)
   where
-    place (!offset, !mvp) itms =
-      let itms' = fmap forEach itms
+    -- the bounding box of the very first item after shifted
+    -- just to have the initial vp value.
+    vp0 = primBoundingBox $ shift offset0 (NE.head (NE.head itmss0))
+
+    place (!offset, !vp) itms =
+      let itms' = fmap (shift offset) itms
           vp' = getLeastUpperBoundingBox itms'
           w' = viewPortWidth vp'
-       in ((offset + w', Just (viewPortSum mvp vp')), itms')
-      where
-        forEach (Primitive shape vp hitEvent) =
-          let shape' = moveShapeBy (offset, 0) shape
-              vp' = moveBoundingBoxBy (offset, 0) vp
-           in Primitive shape' vp' hitEvent
+       in ((offset + w', viewPortSum' vp vp'), itms')
+    shift offset (Primitive shape vp hitEvent) =
+      let shape' = moveShapeBy (offset, 0) shape
+          vp' = moveBoundingBoxBy (offset, 0) vp
+       in Primitive shape' vp' hitEvent
 
-    ((_, mvp1), itms1) = L.mapAccumL place (offset0, Nothing) itms0
+    ((_, vp1), itmss1) = L.mapAccumL place (offset0, vp0) itmss0
 
 -- | place grouped items line by line
 flowLineByLine ::
